Update chunk.py
I am not sure but it seems that the bitwise end should be a bitwise or.
At least this fixes some problems I am encountering when plotting maps.
In particular, when the palette requires 5 bits entries, I am getting an artifact each 13 blocks. So each time the entry is written on two different longs of the state array.